PK-882 library configuration switch for enabling linked accounts
http://marmot.myjetbrains.com/youtrack/issue/PK-882

switch is called allowLinkedAccounts. On Library Admin, it is under the
ILS/Account Integration section. #PK-882

+++ vufind/web/sys/Account/User.php
@@ -203,7 +203,9 @@ class User extends DB_DataObject
 	function getLinkedUsers(){
 		if (is_null($this->linkedUsers)){
 			$this->linkedUsers = array();
-			if ($this->id){
+			/* var Library $library */
+			global $library;
+			if ($this->id && $library->allowLinkedAccounts){
 				require_once ROOT_DIR . '/sys/Account/UserLink.php';
 				$userLink = new UserLink();
 				$userLink->primaryAccountId = $this->id;
@@ -228,7 +230,9 @@ class User extends DB_DataObject
 	function getViewers(){
 		if (is_null($this->viewers)){
 			$this->viewers = array();
-			if ($this->id){
+			/* var Library $library */
+			global $library;
+			if ($this->id && $library->allowLinkedAccounts){
 				require_once ROOT_DIR . '/sys/Account/UserLink.php';
 				$userLink = new UserLink();
 				$userLink->linkedAccountId = $this->id;
@@ -249,34 +253,42 @@ class User extends DB_DataObject
 	 * @param User $user
 	 */
 	function addLinkedUser($user){
-		$linkedUsers = $this->getLinkedUsers();
-		/** @var User $existingUser */
-		foreach ($linkedUsers as $existingUser){
-			if ($existingUser->id == $user->id){
-				//We already have a link to this user
-				return;
+		/* var Library $library */
+		global $library;
+		if ($library->allowLinkedAccounts) {
+			$linkedUsers = $this->getLinkedUsers();
+			/** @var User $existingUser */
+			foreach ($linkedUsers as $existingUser) {
+				if ($existingUser->id == $user->id) {
+					//We already have a link to this user
+					return;
+				}
 			}
+			require_once ROOT_DIR . '/sys/Account/UserLink.php';
+			$userLink                   = new UserLink();
+			$userLink->primaryAccountId = $this->id;
+			$userLink->linkedAccountId  = $user->id;
+			$userLink->insert();
+			$this->linkedUsers[] = clone($user);
 		}
-		require_once ROOT_DIR . '/sys/Account/UserLink.php';
-		$userLink = new UserLink();
-		$userLink->primaryAccountId = $this->id;
-		$userLink->linkedAccountId = $user->id;
-		$userLink->insert();
-		$this->linkedUsers[] = clone($user);
 	}
 
 	function removeLinkedUser($userId){
-		require_once ROOT_DIR . '/sys/Account/UserLink.php';
-		$userLink = new UserLink();
-		$userLink->primaryAccountId = $this->id;
-		$userLink->linkedAccountId = $userId;
-		$ret = $userLink->delete();
+		/* var Library $library */
+		global $library;
+		if ($library->allowLinkedAccounts) {
+			require_once ROOT_DIR . '/sys/Account/UserLink.php';
+			$userLink                   = new UserLink();
+			$userLink->primaryAccountId = $this->id;
+			$userLink->linkedAccountId  = $userId;
+			$ret                        = $userLink->delete();
 
-		//Force a reload of data
-		$this->linkedUsers = null;
-		$this->getLinkedUsers();
+			//Force a reload of data
+			$this->linkedUsers = null;
+			$this->getLinkedUsers();
 
-		return $ret == 1;
+			return $ret == 1;
+		}
 	}
